+++ plugins/forms/dialogs/textarea.js
@@ -113,9 +113,7 @@ CKEDITOR.dialog.add( 'textarea', function( editor ) {
 				'default': '',
 				accessKey: 'Q',
 				value: 'required',
-				setup: function( element ) {
-					this.setValue( element.getAttribute( 'required' ) );
-				},
+				setup: CKEDITOR.plugins.forms._setupRequiredAttribute,
 				commit: function( element ) {
 					if ( this.getValue() )
 						element.setAttribute( 'required', 'required' );